# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Main API",
    "version": "1.0.0",
    "description": "API для работы с пользователями, постами и комментариями"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
    "schemas": {
...
      "ErrorUnauthorized": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "error": {
            "type": "string",
            "example": "Invalid token"
          },
          "statusCode": {
            "type": "integer",
            "example": 401
          }
        }
      },
...
      "ErrorValidation": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          }
        }
      },
      "ErrorNotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Not found"
          },
          "error": {
            "type": "string",
            "example": "Resource not found"
          },
          "statusCode": {
            "type": "integer",
            "example": 404
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clnjak7xj000008l0a9zq3k4f"
          },
          "name": {
            "type": "string",
            "example": "Иван Иванов"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "delivered@resend.dev"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "example": "avatars/clnjak7xj000008l0a9zq3k4f.webp"
          },
          "bio": {
            "type": "string",
            "example": "Люблю разрабатывать API"
          },
          "website": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-01T12:00:00Z"
          }
        }
      },
      "UserUpdateProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "example": "Иван Петров"
          },
          "bio": {
            "type": "string",
            "maxLength": 500,
            "example": "Люблю разрабатывать API"
          },
          "website": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com"
          }
        }
      },
...
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clnjak7xj000008l0a9zq3k4f"
          },
          "title": {
            "type": "string",
            "example": "Как писать хороший API"
          },
          "content": {
            "type": "string",
            "example": "API должен быть простым, предсказуемым и документированным..."
          },
          "authorId": {
            "type": "string",
            "example": "clnja0001user123456789"
          },
          "likesCount": {
            "type": "integer",
            "example": 123
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-01T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-02T14:00:00Z"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/user/profile": {
      "put": {
        "summary": "Обновить профиль текущего пользователя",
        "description": "Изменяет данные профиля текущего авторизованного пользователя",
        "tags": [
          "Users"
        ],
        "operationId": "updateCurrentUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Новые данные профиля",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профиль успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorValidation"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{userId}/posts": {
      "get": {
        "summary": "Получить список постов пользователя с пагинацией",
        "description": "Возвращает список постов указанного пользователя с поддержкой пагинации",
        "tags": [
          "Posts"
        ],
        "operationId": "getUserPosts",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Идентификатор пользователя (cuid)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество записей на странице",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список постов пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Пользователь или посты не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{userId}": {
      "get": {
        "summary": "Получить профиль пользователя по ID",
        "description": "Возвращает данные профиля пользователя по его идентификатору",
        "tags": [
          "Users"
        ],
        "operationId": "getUserProfileById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Идентификатор пользователя (cuid)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      }
    }
  },
...
}